handled a very wacky phageterm-induced corner case.
When phageterm reorganizes a genome to the DTR, it may result
in later CDS detectors (glimmer/prodigal) finding a putative
ORF which crosses the beginning of the contig.  Left unattended,
this may result in shenanigans.  Interestingly, having written
a test for this case, it appears that prodigal-derived features
already take it into account, resulting in a potential partial
duplicate.