Fix error in multi-contig assemblies
There was a problem introduced when I changed how I use phageterm which
caused some assemblies to miss all of the annotations for the first
contig.  Basically, I wasn't (re)setting the feature display_name attribute
and so when it went to merge the annotations it was not able to find the
gene ID to merge into.

I also reorganized a little, moving the merging of annotations to a
later point in the pipeline so that I can add in t(m)RNAs, rho
termination signals, etc.

Finally, I added some logic to interpret the results of phastaf so that,
in the case of multi-contig assemblies, it is able to guess if the
contigs are related or not, and if not it rewrite the assemblies
separately.  The final step of this little piece of logic will therefore
be to invoke the annotation steps on the separate pieces as well as on
the total (in case the guess is wrong).