some pretty extensive changes to how modules are loaded.
I want to have each job run have a clean environment, so I made
some fairly big changes to how modules are loaded.  All module
declarations are now in the function Get_Modules() in Config.pm.

This may get moved to a yml or somesuch file in the future.

In addition, now that I moved one thing to Config.pm, it may be worth
while to put default arguments for each function into it as well.